requirements.txt
# Smart Traffic Analytics - Python Dependencies

# Kafka
confluent-kafka==2.3.0

# MinIO / S3
minio==7.2.3
boto3==1.34.0

# Elasticsearch
elasticsearch==8.11.0

# PostgreSQL
psycopg2-binary==2.9.9
sqlalchemy==2.0.25

# Apache Spark (Batch)
pyspark==3.5.0

# Apache Flink (Streaming)
apache-flink==1.18.0

# Config & Utilities
python-dotenv==1.0.0
requests==2.31.0

# Logging & Monitoring
structlog==24.1.0
